Extract Kernel#format #sprintf to another file
Reduce compiled size (if you don't need this module).
Resolves #1927.
